Rename reducer -> logging
