+++ packages/gravity/kube-testing/src/index.ts
@@ -2,5 +2,5 @@
 // Copyright 2023 DXOS.org
 //
 
-export * from './analysys/reducer';
+export * from './analysys/logging';
 export * from './plan/spec-base';
